codeflix-client/backend-python/requirements.txt
elasticsearch==8.13.2
fastapi[standard]==0.115.4
ipython==8.29.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
confluent-kafka==2.6.1
strawberry-graphql[fastapi]==0.254.0
pytest-mock==3.14.0
PyJWT==2.10.1
cryptography==44.0.0